cppcheck: second batch of fixes
1. Makefile.in - extend the suppression list

Moved the --suppress flags into a CPPCHECK_SUPPRESS variable (the recipe
line was getting unreadable) and added, with a comment block explaining
each:

  postfixOperator          22x `it++` on non-primitive iterators. No
                           measurable cost for the iterator types used
                           here; "fixing" is 22 churn-only edits.

  memsetClassFloat         13x `memset(&x, 0, sizeof(x))` on a struct that
                           contains a float/double. All-zero bytes is a
                           valid 0.0 / 0.0f on every platform ntopng
                           targets (IEEE-754); this is the idiomatic way
                           the codebase zeroes POD-ish structs.

  virtualCallInConstructor  7x  cleanup()/allocateStats()/toggleRxOnlyHost()
  virtualCallInDestructor       /walker() are deliberately called from a
                           ctor/dtor where the *base* implementation is the
                           intended one. Refactoring is invasive and risky.

  nullPointer (path-scoped to src/LuaEngineInterface.cpp,
              src/LuaEngineNetwork.cpp, src/LuaEngineNtop.cpp, src/SNMP.cpp)
                           14x false positives from
                             #define getLuaVMContext(a) (a ? getUserdata(a) : NULL)
                           and getLuaVMUserdata(): cppcheck treats the
                           ": NULL" branch as reachable, but the lua_State*
                           is never NULL inside a Lua C function. Scoped by
                           file so a real null deref elsewhere still shows.

2. Real bugs fixed

src/Utils.cpp  Utils::splitAddressAndVlan()
  The no-'@vlan' branch did `vlan_id = 0;` - assigning the *local pointer*
  parameter instead of `*vlan_id = 0;`. Result: when the address had no
  "@<vlan>" suffix, the caller's vlan_id was left uninitialised (garbage)
  instead of being set to 0. Fixed to `*vlan_id = 0;`.
  (cppcheck: uselessAssignmentPtrArg)

src/Utils.cpp  Utils::parsetime()
  `sscanf(&str[4], "%d%s", &v, what)` with `char what[64]` - unbounded %s,
  stack buffer overflow for a long time spec (these strings come from REST
  parameters such as epoch_begin=now+...). Bounded to "%d%63s".
  (cppcheck: invalidscanf)

src/Utils.cpp  Utils::tokenizer()
  `strchr(arg, c)` was called before the function's own `if (arg)` guard a
  few lines down - inconsistent, and a crash if a caller ever passes NULL
  (which the later guard shows was considered possible). Added an
  arg == NULL early return (sets *data = NULL, returns NULL) at the top.
  (cppcheck: nullPointerRedundantCheck)

src/IpAddress.cpp  IpAddress::isLocalInterfaceAddress() (broadcast calc)
  `~((1 << (32 - nmask_bits)) - 1)` with nmask_bits in 1..30 -> the shift
  count reaches 31, i.e. `1 << 31` on a signed int = shift into the sign
  bit, which is implementation-defined / UB. Changed the literal to
  `(u_int32_t)1`. `nmask` is already u_int32_t.
  (cppcheck: shiftTooManyBitsSigned)

src/HTTPserver.cpp (x2), src/Ntop.cpp (x5), src/LuaEngineNtop.cpp (x1)
  `mg_md5(buf, a, b, NULL)` - mg_md5() is variadic and stops at a NULL
  string pointer, but a bare `NULL` may be passed as `int 0` (only
  well-defined when NULL is a pointer constant of the right width). Passing
  `(char *) NULL` makes the sentinel explicitly pointer-typed. No
  behavioural change on LP64 Linux/macOS; removes the UB.
  (cppcheck: varFuncNullUB)

include/Flow.h  Flow::setAddressFamilyProtocol()
  `ndpiAddressFamilyProtocol = strdup(proto);` with no free of the previous
  value and no NULL check on `proto`. The setter can be called more than
  once per flow as detection progresses -> each subsequent call leaked the
  previous strdup() (the destructor only frees the last one). Now frees the
  old buffer first and guards `proto == NULL`.
  (cppcheck: publicAllocationError)

src/ParsedFlowCore.cpp  both constructors
  Neither the default ctor nor the copy ctor initialised next_hop,
  mapped_next_hop and mapped_exporter_device_ip (struct ndpi_in6_addr, used
  for collected/SNMP-mapped next-hop and exporter addresses). Default ctor
  now memset()s them; copy ctor now memcpy()s them from the source.
  (cppcheck: uninitMemberVar / missingMemberCopy)

src/GenericTrafficElement.cpp  bytes_thpt_diff
  `float bytes_thpt_diff` was never initialised in resetStats() (called by
  the default ctor) nor copied by the copy ctor. Initialised to 0 in
  resetStats() and copied in the copy ctor. (This member currently has no
  other reader in the tree, so the effect is limited, but the copy ctor was
  literally reading an uninitialised float.)
  (cppcheck: uninitMemberVar / missingMemberCopy)

src/Host.cpp  Host::triggerAlert()
  `u_int64_t alert_rowid = alert->getRowID();` ran *before*
  `if (alert == NULL) return false;` - the null check right after it is
  pointless if the deref already happened. Moved the assignment below the
  check.
  (cppcheck: nullPointerRedundantCheck)

src/Redis.cpp  Redis::lrpop()
  After `*buf = strdup(reply->str);` the code did `if (buf) rc = 0; else
  rc = -2;` - `buf` is the `char **` parameter and is always non-NULL, so
  the strdup-failure branch (rc = -2) was dead. Changed to `if (*buf)`.
  (cppcheck: nullPointerRedundantCheck)

src/Ntop.cpp  Ntop::changeAllowedIfname()
  `strstr(allowed_ifname, ":__")` at the top of the function, while the
  body a few lines later checks `allowed_ifname != NULL && allowed_ifname[0]`
  - inconsistent, and a crash if allowed_ifname is NULL. Reordered so the
  username guard runs first, then the strstr is done only when
  allowed_ifname is non-NULL.
  (cppcheck: nullPointerRedundantCheck)

+++ src/IpAddress.cpp
@@ -176,7 +176,7 @@ void IpAddress::checkIP() {
       if (!ntop->isBroadcastIPDisabled() && nmask_bits > 0 &&
           nmask_bits < 31) { /* /0 no mask, /32 is just an host, /31 is a
                                 point-to-point */
-        nmask = ~((1 << (32 - nmask_bits)) - 1);
+        nmask = ~(((u_int32_t)1 << (32 - nmask_bits)) - 1);
         if (a == (a |
                   ~nmask) /* e.g., 10.0.0.0/8 -> matches 10.255.255.255.255 */
             || a == (a & nmask) /* e.g., 10.0.0.0/8 -> matches 10.0.0.0 */)
